Test code in place, wip
