+++ server/routes/post.routes.js
@@ -6,13 +6,13 @@ const router = Router();
 // Get all Posts
 router.route('/posts').get(PostController.getPosts);
 
-// Get one post by _id
-router.route('/posts/:_id').get(PostController.getPost);
+// Get one post by slug
+router.route('/posts/:slug').get(PostController.getPost);
 
 // Add a new Post
 router.route('/posts').post(PostController.addPost);
 
-// Delete a post by _id
-router.route('/posts/:_id').delete(PostController.deletePost);
+// Delete a post by slug
+router.route('/posts/:slug').delete(PostController.deletePost);
 
 module.exports = router;
